@@ -2,7 +2,7 @@ source 'https://rubygems.org'
 
 ruby '2.4.1'
 
-gem 'rails', '~> 4.2.1'
+gem 'rails', '~> 5.1.0'
 gem 'pg'
 gem 'jquery-rails'
 gem 'jquery-ui-rails'
@@ -22,7 +22,6 @@ gem 'mailgun-ruby', require: 'mailgun'
 gem 'httparty'
 
 group :development do
-  gem 'quiet_assets'
   gem 'better_errors'
   gem 'binding_of_caller'
   gem 'letter_opener'
