@@ -1,5 +1,5 @@
-# Load the rails application
-require File.expand_path('../application', __FILE__)
+# Load the Rails application.
+require_relative 'application'
 
-# Initialize the rails application
-Textbook::Application.initialize!
+# Initialize the Rails application.
+Rails.application.initialize!
